[STF] Move Constants to STF-Self-Tests
