+++ stf.test/test/saros/stf/test/consistency/EditDuringInvitationStressTest.java
@@ -15,7 +15,7 @@ import saros.stf.client.StfTestCase;
 import saros.stf.client.tester.SarosTester;
 import saros.stf.client.util.Util;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 import saros.test.util.EclipseTestThread;
 
 public class EditDuringInvitationStressTest extends StfTestCase {
