+++ stf.test/test/saros/stf/test/consistency/EditDuringInvitationTest.java
@@ -13,7 +13,7 @@ import org.junit.Test;
 import saros.stf.annotation.TestLink;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 import saros.test.util.EclipseTestThread;
 
 @TestLink(id = "Saros-36_edit_during_invitation")
