+++ stf.test/test/saros/stf/test/consistency/EditDuringNonHostInvitationTest.java
@@ -13,7 +13,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 import saros.test.util.EclipseTestThread;
 
 public class EditDuringNonHostInvitationTest extends StfTestCase {
