+++ stf.test/test/saros/stf/test/consistency/ModifyDocumentBeforeProjectNegotiationTest.java
@@ -12,7 +12,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class ModifyDocumentBeforeProjectNegotiationTest extends StfTestCase {
