+++ stf.test/test/saros/stf/test/filefolderoperations/FileOperationsTest.java
@@ -15,7 +15,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class FileOperationsTest extends StfTestCase {
