+++ stf.test/test/saros/stf/test/invitation/ParallelInvitationWithTerminationByHostTest.java
@@ -14,7 +14,7 @@ import org.eclipse.core.runtime.CoreException;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class ParallelInvitationWithTerminationByHostTest extends StfTestCase {
