+++ stf.test/test/saros/stf/test/invitation/Share2UsersSequentiallyTest.java
@@ -11,7 +11,7 @@ import saros.session.User.Permission;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class Share2UsersSequentiallyTest extends StfTestCase {
