+++ stf/test/saros/stf/test/stf/basicwidget/BasicWidgetTreeTest.java
@@ -42,7 +42,7 @@ import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.server.rmi.remotebot.widget.IRemoteBotShell;
 import saros.stf.server.rmi.remotebot.widget.IRemoteBotView;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class BasicWidgetTreeTest extends StfTestCase {
