+++ stf/test/saros/stf/test/stf/contextmenu/ContextMenuOpenTest.java
@@ -11,7 +11,7 @@ import org.junit.After;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class ContextMenuOpenTest extends StfTestCase {
