+++ stf/test/saros/stf/test/stf/contextmenu/ContextMenuShareWithTest.java
@@ -8,7 +8,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class ContextMenuShareWithTest extends StfTestCase {
