+++ stf/test/saros/stf/test/stf/editor/EditorWithoutSessionTest.java
@@ -8,7 +8,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class EditorWithoutSessionTest extends StfTestCase {
