+++ stf/test/saros/stf/test/stf/menubar/MenuRefactorTest.java
@@ -7,7 +7,7 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class MenuRefactorTest extends StfTestCase {
