+++ stf/test/saros/stf/test/stf/menubar/MenuSarosByAliceBobCarlTest.java
@@ -11,7 +11,7 @@ import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class MenuSarosByAliceBobCarlTest extends StfTestCase {
