+++ stf/test/saros/stf/test/stf/menubar/MenuSarosByAliceBobTest.java
@@ -20,7 +20,7 @@ import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
 import saros.stf.server.rmi.remotebot.widget.IRemoteBotShell;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class MenuSarosByAliceBobTest extends StfTestCase {
