+++ stf/test/saros/stf/test/stf/menubar/MenuTest.java
@@ -10,7 +10,7 @@ import org.junit.After;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class MenuTest extends StfTestCase {
