+++ stf/test/saros/stf/test/stf/menubar/SarosPreferencesTest.java
@@ -23,7 +23,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.server.rmi.remotebot.widget.IRemoteBotShell;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class SarosPreferencesTest extends StfTestCase {
