+++ stf/test/saros/stf/test/stf/view/sarosview/content/ContactsByAliceBobTest.java
@@ -12,7 +12,7 @@ import org.junit.Test;
 import saros.stf.client.StfTestCase;
 import saros.stf.client.util.Util;
 import saros.stf.shared.Constants.TypeOfCreateProject;
-import saros.stf.test.Constants;
+import saros.stf.test.stf.Constants;
 
 public class ContactsByAliceBobTest extends StfTestCase {
